+++ src/interpolate.rs
@@ -1,23 +1,23 @@
 use crate::Color;
 
-fn basis(colors: &[Color], component: fn(&Color) -> u8, t: f64) -> u8 {
+fn basis(colors: &[Color], component: fn(&Color) -> u8, t: f32) -> u8 {
     let n = colors.len() - 1;
-    let i = ((t * n as f64).floor() as usize).min(n - 1);
+    let i = ((t * n as f32).floor() as usize).min(n - 1);
 
-    let v1 = component(&colors[i]) as f64;
-    let v2 = component(&colors[i + 1]) as f64;
+    let v1 = component(&colors[i]) as f32;
+    let v2 = component(&colors[i + 1]) as f32;
     let v0 = if i > 0 {
-        component(&colors[i - 1]) as f64
+        component(&colors[i - 1]) as f32
     } else {
         2.0 * v1 - v2
     };
     let v3 = if i < n - 1 {
-        component(&colors[i + 2]) as f64
+        component(&colors[i + 2]) as f32
     } else {
         2.0 * v2 - v1
     };
 
-    let t1 = (t - i as f64 / n as f64) * n as f64;
+    let t1 = (t - i as f32 / n as f32) * n as f32;
     let t2 = t1 * t1;
     let t3 = t2 * t1;
 
@@ -29,7 +29,7 @@ fn basis(colors: &[Color], component: fn(&Color) -> u8, t: f64) -> u8 {
 }
 
 // 0.0 <= t <= 1.0
-pub fn spline(colors: &[Color], t: f64) -> Color {
+pub fn spline(colors: &[Color], t: f32) -> Color {
     Color {
         r: basis(colors, |c| c.r, t),
         g: basis(colors, |c| c.g, t),
